backend/pyproject.toml
[project]
name = "pans-software-backend"
version = "0.1.0"
description = "Business-grade FastAPI backend for the Pans Software take-home assessment."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0,<1.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "langchain-openai>=0.3.0,<1.0.0",
    "langgraph>=0.3.0,<1.0.0",
    "lancedb>=0.29.2,<1.0.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "python-multipart>=0.0.20,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "uvicorn[standard]>=0.41.0,<1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1,<1.0.0",
    "mypy>=1.19.1,<2.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.0.0,<2.0.0",
    "ruff>=0.15.4,<1.0.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
markers = [
    "live_e2e: real end-to-end tests that use live OpenAI and LanceDB",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["app", "tests"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
